consumer/requirements.txt
delta-spark==3.1.0
pyspark==3.5.1python-dotenv
